rfct: Move MorseDialog to seperate file
